app/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart
pydantic
pydantic-core
pillow
opencv-python
pytesseract
python-dateutil
requests
scikit-learn
numpy
scipy
transformers
torch
onnx==1.16.0
onnxruntime==1.19.2
pandas
PyMySQL